将T-Circle-S3与T-CamerPlus-S3的codec放入boards下
