+++ main/CMakeLists.txt
@@ -4,8 +4,6 @@ set(SOURCES "audio_codecs/audio_codec.cc"
             "audio_codecs/es8311_audio_codec.cc"
             "audio_codecs/es8388_audio_codec.cc"
             "audio_codecs/cores3_audio_codec.cc"
-            "audio_codecs/tcircles3_audio_codec.cc"
-            "audio_codecs/tcamerapluss3_audio_codec.cc"
             "led/single_led.cc"
             "led/circular_strip.cc"
             "display/display.cc"
