+++ main/boards/lilygo-t-cameraplus-s3/tcamerapluss3_audio_codec.h
@@ -1,12 +1,12 @@
 #ifndef _TCIRCLES3_AUDIO_CODEC_H
 #define _TCIRCLES3_AUDIO_CODEC_H
 
-#include "audio_codec.h"
+#include "audio_codecs/audio_codec.h"
 
 #include <esp_codec_dev.h>
 #include <esp_codec_dev_defaults.h>
 
-class Tcircles3AudioCodec : public AudioCodec {
+class Tcamerapluss3AudioCodec : public AudioCodec {
 private:
     const audio_codec_data_if_t *data_if_ = nullptr;
     const audio_codec_ctrl_if_t *out_ctrl_if_ = nullptr;
@@ -23,11 +23,11 @@ private:
     virtual int Write(const int16_t *data, int samples) override;
 
 public:
-    Tcircles3AudioCodec(int input_sample_rate, int output_sample_rate,
+    Tcamerapluss3AudioCodec(int input_sample_rate, int output_sample_rate,
         gpio_num_t mic_bclk, gpio_num_t mic_ws, gpio_num_t mic_data,
         gpio_num_t spkr_bclk, gpio_num_t spkr_lrclk, gpio_num_t spkr_data,
         bool input_reference);
-    virtual ~Tcircles3AudioCodec();
+    virtual ~Tcamerapluss3AudioCodec();
 
     virtual void SetOutputVolume(int volume) override;
     virtual void EnableInput(bool enable) override;
