+++ main/boards/lilygo-t-circle-s3/tcircles3_audio_codec.cc
@@ -132,7 +132,7 @@ int Tcircles3AudioCodec::Read(int16_t *dest, int samples){
     return samples;
 }
 
-void AdjustVolume(const int16_t *input_data, int16_t *output_data, size_t samples, float volume){
+void AdjustTcircles3Volume(const int16_t *input_data, int16_t *output_data, size_t samples, float volume){
     for (size_t i = 0; i < samples; i++){
         output_data[i] = (float)input_data[i] * volume;
     }
@@ -142,7 +142,7 @@ int Tcircles3AudioCodec::Write(const int16_t *data, int samples){
     if (output_enabled_){
         size_t bytes_read;
         auto output_data = (int16_t *)malloc(samples * sizeof(int16_t));
-        AdjustVolume(data, output_data, samples, (float)(volume_ / 100.0));
+        AdjustTcircles3Volume(data, output_data, samples, (float)(volume_ / 100.0));
         i2s_channel_write(tx_handle_, output_data, samples * sizeof(int16_t), &bytes_read, portMAX_DELAY);
         free(output_data);
     }
